Refactor:  캘린더 관련 상수분리 #17
