+++ src/constants/anniversary.constants.ts
@@ -1,0 +1,2 @@
+// test용 상수 - 백엔드 연동 후 삭제 예정
+export const STORAGE_KEY = 'custom_anniversaries';
