+++ src/constants/calendar.constants.ts
@@ -1,0 +1,4 @@
+// 백엔드 연동 시 수정 - 백엔드 API에서 공휴일 데이터를 정제하여 제공하면 이 상수 삭제 가능
+export const CALENDAR_ID = 'ko.south_korea#holiday@group.v.calendar.google.com';
+
+export const WEEKDAYS = ['일', '월', '화', '수', '목', '금', '토'];
